[ICU-790] Removed the ULONG_MAX reference, which should never have been there. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/3330

+++ icu4c/source/test/cintltst/creststn.c
@@ -61,7 +61,7 @@ randul()
 static double
 randd()
 {
-    return ((double)randul()) / ULONG_MAX;
+    return ((double)randul()) / UINT32_MAX;
 }
 
 /**
